Refactor get_miller_index to Lattice
