+++ pymatgen/core/lattice.py
@@ -5,6 +5,17 @@
 from __future__ import division, unicode_literals
 import math
 import itertools
+import warnings
+
+from functools import reduce
+try:
+    # New Py>=3.5 import
+    from math import gcd
+except ImportError:
+    # Deprecated import from Py3.5 onwards.
+    from fractions import gcd
+
+from fractions import Fraction
 
 from six.moves import map, zip
 
@@ -30,7 +41,6 @@ __status__ = "Production"
 __date__ = "Sep 23, 2011"
 
 
-
 class Lattice(MSONable):
     """
     A lattice object.  Essentially a matrix with conversion matrices. In
@@ -39,7 +49,6 @@ class Lattice(MSONable):
     """
 
     # Properties lazily generated for efficiency.
-
 
     def __init__(self, matrix):
         """
@@ -1117,3 +1126,93 @@ class Lattice(MSONable):
         mapped_vec = self.get_cartesian_coords(jimage + frac_coords2
                                                - frac_coords1)
         return np.linalg.norm(mapped_vec), jimage
+
+    def get_miller_index_from_sites(self, coords, coords_are_cartesian=True,
+                                    round_dp=4, verbose=True):
+        """
+        Get the Miller index of a plane from a set of sites.
+
+        A minimum of 3 sites are required. If more than 3 sites are given
+        the best plane that minimises the distance to all points will be
+        calculated.
+
+        Args:
+            coords (iterable): A list or numpy array of site coordinates. Can be
+                cartesian or fractional coordinates. If more than three sites
+                are provided, the best plane that minimises the distance to all
+                sites will be calculated.
+            coords_are_cartesian (bool, optional): Whether the coordinates are
+                in cartesian space. If using fractional coordinates set to
+                False.
+            round_dp (int, optional): The number of decimal places to round the
+                miller index to.
+            verbose (bool, optional): Whether to print warnings.
+
+        Returns:
+            (tuple): The Miller index.
+        """
+        if coords_are_cartesian:
+            coords = [self.get_fractional_coords(c) for c in coords]
+
+        coords = np.asarray(coords)
+        g = coords.sum(axis=0) / coords.shape[0]
+
+        # run singular value decomposition
+        _, _, vh = np.linalg.svd(coords - g)
+
+        # get unitary normal vector
+        u_norm = vh[2, :]
+        return get_integer_index(u_norm, round_dp=round_dp, verbose=verbose)
+
+
+def get_integer_index(miller_index, round_dp=4, verbose=True):
+    """
+    Attempt to convert a vector of floats to whole numbers.
+
+    Args:
+        miller_index (list of float): A list miller indexes.
+        round_dp (int, optional): The number of decimal places to round the
+            miller index to.
+        verbose (bool, optional): Whether to print warnings.
+
+    Returns:
+        (tuple): The Miller index.
+    """
+    miller_index = np.asarray(miller_index)
+
+    # deal with the case we have small irregular floats
+    # that are all equal or factors of each other
+    miller_index /= min([m for m in miller_index if m != 0])
+    miller_index /= np.max(np.abs(miller_index))
+
+    # deal with the case we have nice fractions
+    md = [Fraction(n).limit_denominator(12).denominator for n in miller_index]
+    miller_index *= reduce(lambda x, y: x * y, md)
+    round_miller_index = np.int_(np.round(miller_index, 1))
+    miller_index /= np.abs(reduce(gcd, round_miller_index))
+
+    # round to a reasonable precision
+    miller_index = np.array([round(h, round_dp) for h in miller_index])
+
+    # need to recalculate this after rounding as values may have changed
+    round_miller_index = np.int_(np.round(miller_index, 1))
+    if (np.any(np.abs(miller_index - round_miller_index) > 1e-6) and
+            verbose):
+        warnings.warn("Non-integer encountered in Miller index")
+
+    # minimise the number of negative indexes
+    miller_index += 0  # converts -0 to 0
+
+    def n_minus(index):
+        return len([h for h in index if h < 0])
+
+    if n_minus(miller_index) > n_minus(miller_index * -1):
+        miller_index *= -1
+
+    # if only one index is negative, make sure it is the smallest
+    # e.g. (-2 1 0) -> (2 -1 0)
+    if (sum(miller_index != 0) == 2 and n_minus(miller_index) == 1
+            and abs(min(miller_index)) > max(miller_index)):
+        miller_index *= -1
+
+    return tuple(miller_index)
