+++ pymatgen/core/surface.py
@@ -17,7 +17,6 @@ import warnings
 import copy
 import os
 import json
-from fractions import Fraction
 
 import numpy as np
 from scipy.spatial.distance import squareform
@@ -1680,59 +1679,6 @@ def generate_all_slabs(structure, max_index, min_slab_size, min_vacuum_size,
     return all_slabs
 
 
-def get_integer_index(miller_index, round_dp=4, verbose=True):
-    """
-    Attempt to convert a vector of floats to whole numbers.
-
-    Args:
-        miller_index (list of float): A list miller indexes.
-        round_dp (int, optional): The number of decimal places to round the
-            miller index to.
-        verbose (bool, optional): Whether to print warnings.
-
-    Returns:
-        (tuple): The Miller index.
-    """
-    miller_index = np.asarray(miller_index)
-
-    # deal with the case we have small irregular floats
-    # that are all equal or factors of each other
-    miller_index /= min([m for m in miller_index if m != 0])
-    miller_index /= np.max(np.abs(miller_index))
-
-    # deal with the case we have nice fractions
-    md = [Fraction(n).limit_denominator(12).denominator for n in miller_index]
-    miller_index *= reduce(lambda x, y: x * y, md)
-    round_miller_index = np.int_(np.round(miller_index, 1))
-    miller_index /= np.abs(reduce(gcd, round_miller_index))
-
-    # round to a reasonable precision
-    miller_index = np.array([round(h, round_dp) for h in miller_index])
-
-    # need to recalculate this after rounding as values may have changed
-    round_miller_index = np.int_(np.round(miller_index, 1))
-    if (np.any(np.abs(miller_index - round_miller_index) > 1e-6) and
-            verbose):
-        warnings.warn("Non-integer encountered in Miller index")
-
-    # minimise the number of negative indexes
-    miller_index += 0  # converts -0 to 0
-
-    def n_minus(index):
-        return len([h for h in index if h < 0])
-
-    if n_minus(miller_index) > n_minus(miller_index * -1):
-        miller_index *= -1
-
-    # if only one index is negative, make sure it is the smallest
-    # e.g. (-2 1 0) -> (2 -1 0)
-    if (sum(miller_index != 0) == 2 and n_minus(miller_index) == 1
-            and abs(min(miller_index)) > max(miller_index)):
-        miller_index *= -1
-
-    return tuple(miller_index)
-
-
 def miller_index_from_sites(lattice, coords, coords_are_cartesian=True,
                             round_dp=4, verbose=True):
     """
@@ -1763,18 +1709,9 @@ def miller_index_from_sites(lattice, coords, coords_are_cartesian=True,
     Returns:
         (tuple): The Miller index.
     """
-    if coords_are_cartesian:
-        coords = [lattice.get_fractional_coords(c) for c in coords]
-
-    coords = np.asarray(coords)
-    G = coords.sum(axis=0) / coords.shape[0]
-
-    # run singular value decomposition
-    u, s, vh = np.linalg.svd(coords - G)
-
-    # get unitary normal vector
-    u_norm = vh[2, :]
-    return get_integer_index(u_norm, round_dp=round_dp, verbose=verbose)
+    return lattice.get_miller_index_from_sites(
+        coords, coords_are_cartesian=coords_are_cartesian, round_dp=round_dp,
+        verbose=verbose)
 
 
 def center_slab(slab):
